move eager connet to userProvider
